Coerce performance fields to float

+++ visigoth/experiment.py
@@ -217,9 +217,9 @@ class Experiment(object):
         if self.trial_data:
             data = pd.DataFrame(self.trial_data)
             if "correct" in data:
-                mean_acc = data["correct"].mean()
+                mean_acc = data["correct"].astype(float).mean()
             if "rt" in data:
-                mean_rt = data["rt"].mean()
+                mean_rt = data["rt"].astype(float).mean()
         return mean_acc, mean_rt
 
     def show_performance(self, mean_acc, mean_rt):
